feat(order): add orders endpoints and requests

+++ server/migrations/20250112113756_create_orders_table.js
@@ -2,7 +2,7 @@ export async function up(knex) {
   await knex.schema.createTable("orders", function (table) {
     table.increments("id").primary();
     table.string("info").notNullable();
-    table.integer("weight").notNullable();
+    table.float("weight").notNullable();
     table.string("from").notNullable();
     table.string("to").notNullable();
     table.timestamp("create_at").defaultTo(knex.fn.now());
